Reorganize server package structure and introduce HTTPBulkServer
- Moved `WebSocketDeltaChannel` and `WebSocketDeltaServer` to a new `server.delta` package.
- Added `HTTPBulkServer` class with basic start/stop methods.
- Refactored LionWebServer to configure `wsPort`, `httpPort`, and `webUIPort` independently.
- Updated related imports and adjusted CLI options for enhanced clarity.
