+++ server/src/main/kotlin/io/lionweb/server/bulk/HTTPBulkServer.kt
@@ -1,0 +1,13 @@
+package io.lionweb.server.bulk
+
+import io.lionweb.client.inmemory.InMemoryServer
+
+class HTTPBulkServer(val inMemoryServer: InMemoryServer, val port: Int) {
+    fun start() {
+
+    }
+
+    fun stop() {
+
+    }
+}
